+++ classes/Conf/Workflows/class.xoctWorkflowTableGUI.php
@@ -91,7 +91,7 @@ class xoctWorkflowTableGUI extends TableGUI
      * @throws DICException
      * @throws ilTemplateException
      */
-    protected function getColumnValue($column, $row, $format = self::DEFAULT_FORMAT)
+    protected function getColumnValue(string $column, /*array*/ $row, int $format = self::DEFAULT_FORMAT) : string
     {
         switch ($column) {
             case 'id':
@@ -135,7 +135,7 @@ class xoctWorkflowTableGUI extends TableGUI
      * @inheritDoc
      * @throws DICException
      */
-    protected function getSelectableColumns2()
+    protected function getSelectableColumns2() : array
     {
         return [
             ['txt' => self::dic()->language()->txt('id'), 'id' => 'id'],